+++ zebedee-cms/src/test/java/com/github/onsdigital/zebedee/model/approval/task/CollectionPdfGeneratorTest.java
@@ -1,5 +1,6 @@
 package com.github.onsdigital.zebedee.model.approval.task;
 
+import com.github.onsdigital.zebedee.content.base.ContentLanguage;
 import com.github.onsdigital.zebedee.content.page.base.PageType;
 import com.github.onsdigital.zebedee.exceptions.NotFoundException;
 import com.github.onsdigital.zebedee.exceptions.ZebedeeException;
@@ -9,7 +10,6 @@ import com.github.onsdigital.zebedee.model.ContentWriter;
 import com.github.onsdigital.zebedee.model.approval.tasks.CollectionPdfGenerator;
 import com.github.onsdigital.zebedee.reader.ContentReader;
 import com.github.onsdigital.zebedee.reader.Resource;
-import com.github.onsdigital.zebedee.reader.data.language.ContentLanguage;
 import com.github.onsdigital.zebedee.service.PdfService;
 import org.junit.Before;
 import org.junit.Test;
