+++ zebedee-reader/src/main/java/com/github/onsdigital/zebedee/reader/CompositeContentReader.java
@@ -4,7 +4,7 @@ import com.github.onsdigital.zebedee.content.dynamic.browse.ContentNode;
 import com.github.onsdigital.zebedee.content.page.base.Page;
 import com.github.onsdigital.zebedee.exceptions.BadRequestException;
 import com.github.onsdigital.zebedee.exceptions.ZebedeeException;
-import com.github.onsdigital.zebedee.reader.data.language.ContentLanguage;
+import com.github.onsdigital.zebedee.content.base.ContentLanguage;
 
 import java.io.IOException;
 import java.net.URI;
