+++ zebedee-reader/src/main/java/com/github/onsdigital/zebedee/reader/api/endpoint/ResolveDatasets.java
@@ -9,9 +9,9 @@ import com.github.onsdigital.zebedee.content.page.taxonomy.ProductPage;
 import com.github.onsdigital.zebedee.exceptions.BadRequestException;
 import com.github.onsdigital.zebedee.exceptions.NotFoundException;
 import com.github.onsdigital.zebedee.exceptions.ZebedeeException;
+import com.github.onsdigital.zebedee.content.base.ContentLanguage;
 import com.github.onsdigital.zebedee.reader.api.ReadRequestHandler;
 import com.github.onsdigital.zebedee.reader.api.bean.DatasetSummary;
-import com.github.onsdigital.zebedee.reader.data.language.ContentLanguage;
 import com.github.onsdigital.zebedee.reader.resolver.DatasetSummaryResolver;
 import com.github.onsdigital.zebedee.reader.util.ReaderResponseWriter;
 import com.github.onsdigital.zebedee.reader.util.ResponseWriter;
